fix: Fixed issue in loading disqus
Due to configurations on the disqus site, the api wont load locally, and the api requests doesnt alert user.

I therefore added a new disqus site without URL configs, so it should load locally, and removed identifier configs.

Also wrapped the disqus script in an exceptionhandler, so it will alert the user upon errors, not just write it in the modal.

Finally did some minor non functional fixes in comments and an alert on foursquare request failure